+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/controller/TestController.java
@@ -5,6 +5,8 @@ import hamburger.fashiontoday.domain.lookitem.LookitemRepository;
 import hamburger.fashiontoday.domain.lookitemclass.LookItemClassRepository;
 import hamburger.fashiontoday.domain.lookitemclass.LookitemClass;
 import hamburger.fashiontoday.domain.lookstructure.LookStructureRepository;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,6 +22,9 @@ import org.springframework.web.bind.annotation.RestController;
  */
 @RestController
 public class TestController {
+
+    //
+    private static Logger logger = LogManager.getLogger(TestController.class);
 
     @Autowired
     LookitemRepository lookItemRepository;
