+++ FashionToday_Back/src/main/java/hamburger/fashiontoday/controller/UserController.java
@@ -1,0 +1,49 @@
+package hamburger.fashiontoday.controller;
+
+import hamburger.fashiontoday.domain.member.Member;
+import hamburger.fashiontoday.domain.member.MemberRepository;
+import hamburger.fashiontoday.dto.UserInfo;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+
+/**
+ * @프로그램ID : HAM-PB-1002-J
+ * @프로그램명 : UserController.java
+ * @author : 심기성
+ * @date : 2019.09.29
+ * @version : 0.5
+ *
+ */
+@RestController
+@CrossOrigin(origins = "8080")
+public class UserController {
+
+    // 로그를 찍기 위한 Logger
+    private static Logger logger = LogManager.getLogger(UserController.class);
+
+    // 프로그램 아이디 및 에러 코드
+    private String programId = "HAM-PB-4002-J";
+    private String errorCode = "";
+
+    // Member객체를 관리하는 memberrepository입니다.
+    @Autowired
+    MemberRepository memberRepository;
+
+    /**
+     *
+     * @return
+     */
+    @PostMapping("/userInfo")
+    public UserInfo userInfo(){
+        Member member = memberRepository.findByMId(3);
+        UserInfo userInfo = new UserInfo(member.getMId(),member.getMName(),member.getMStar(),member.getMProfileUrl(),10,"200");
+        logger.debug(programId + " : userInfo - success : memberId = "+member.getMId());
+        return userInfo;
+    }
+
+
+
+}
